some change6

+++ entrypoint.sh
@@ -1,10 +1,5 @@
 #!/bin/sh
 
-# Install Dependencies
-composer install --optimize-autoloader --no-dev
-composer dump-autoload
-composer require fakerphp/faker --dev
-
 # Clear Cache
 php artisan optimize:clear
 php artisan cache:clear
@@ -17,8 +12,12 @@ if [ ! -L "public/storage" ]; then
     php artisan storage:link
 fi
 
-# Run Migrations & Seeding
-php artisan migrate:fresh --seed --force
+# Run Migrations & Seeding (Only in local/dev)
+if [ "$APP_ENV" != "production" ]; then
+    php artisan migrate:fresh --seed --force
+else
+    php artisan migrate --force
+fi
 
 # Cache Config & Routes for Performance
 php artisan config:cache
